fix: Modifique que en el formulario los select se vieran mas claritos cuando esta deshabilitado, ajuste el nombre del componente a profile y el icono del calendario a mas clarito y fuera del input
