--- frontend/src/pages/MedicalProfilePage.tsx
@@ -1,9 +1,0 @@
-import { MedicalProfile } from '../components/MedicalProfile/MedicalProfile.tsx';
-
-export const MedicalProfilePage = () => {
-  return (
-    <main className="flex min-h-screen items-center justify-center p-4">
-      < MedicalProfile/>
-    </main>
-  );
-};
