+++ frontend/src/pages/ProfilePage.tsx
@@ -1,0 +1,9 @@
+import { Profile } from '../components/Profile/Profile.tsx';
+
+export const ProfilePage = () => {
+  return (
+    <main className="flex min-h-screen items-center justify-center p-4">
+      < Profile/>
+    </main>
+  );
+};
